feat(telegram_bot): add support for custom telegram bot API URL
This commit introduces the ability to configure a custom base URL
for the Telegram Bot API. This is useful for self-hosted bot API
servers. The base URL can be configured in the `config.toml` file.
If a base URL is provided, the application will use it to
communicate with the Telegram Bot API.

Signed-off-by: Troublescope <121922135+troublescope@users.noreply.github.com>